Fix OpenCV integration.
Attempts to load the OpenCV library failed because it could not find the
JNI shared library. For some reason you need to place it in libs/ARCH
and it doesn't pull it from configured directory (or maybe that wasn't
done right). It also seems external building wasn't working.

Copying the JNI lib into the libs/ARCH folder allows the library to be
loaded and a Mat created/manipulated. Testing the clustering is still
TBD.